ADDED schema for Cascade Rules of Patient, Claims, and User's onDelete reference, ADDED hospitalId in comments, Self referencing hospitalId for HOSPITAL_MANAGER roles for  manager's chats, UPDATED comments create, and listing for posting and fetching manager level comments, ADDED API to fetch chat list for ADMIN side UI, ADDED Patient delete if no claims are associated and Claim delete if status is DRAFT on Delete API

+++ src/patients/patients.service.ts
@@ -97,6 +97,8 @@ export class PatientsService {
     id: string; 
     hospitalUserId: string 
   }): Promise<Patient>{
+    const claimsCount = await this.prisma.insuranceRequest.count({ where: { patientId: where.id } });
+    if (claimsCount>0) throw new Error("Cannot delete patient with existing claims");
     return await this.prisma.patient.delete({
       where
     })
